requirements.txt
Django~=4.2
mysqlclient~=2.1.1
django-cors-headers~=3.14.0
djangorestframework~=3.14.0
pytz~=2023.3